requirements.txt
numpy>=1.20
scipy>=1.7
matplotlib>=3.4
streamlit>=1.28
openai>=1.0.0
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
PyYAML>=6.0
